chore(member): 사용자 관련 클래스 패키징
