+++ src/main/java/roomescape/controller/response/ReservationResponse.java
@@ -1,6 +1,6 @@
 package roomescape.controller.response;
 
-import roomescape.model.Member;
+import roomescape.model.member.Member;
 import roomescape.model.Reservation;
 import roomescape.model.ReservationTime;
 import roomescape.model.Theme;
