+++ src/main/java/roomescape/model/member/Member.java
@@ -1,4 +1,4 @@
-package roomescape.model;
+package roomescape.model.member;
 
 import java.util.Objects;
