+++ src/main/java/roomescape/model/member/Name.java
@@ -1,10 +1,10 @@
-package roomescape.model;
+package roomescape.model.member;
 
-public class Password {
+public class Name {
 
     private final String value;
 
-    public Password(String value) {
+    public Name(String value) {
         validate(value);
         this.value = value;
     }
@@ -15,7 +15,7 @@ public class Password {
 
     private void validateNull(String value) {
         if (value == null || value.isBlank()) {
-            throw new IllegalStateException("비밀번호는 null 혹은 빈 문자열일 수 없습니다.");
+            throw new IllegalStateException("사용자 이름은 null 혹은 빈 문자열일 수 없습니다.");
         }
     }
