+++ src/main/java/roomescape/model/member/Password.java
@@ -1,10 +1,10 @@
-package roomescape.model;
+package roomescape.model.member;
 
-public class Name {
+public class Password {
 
     private final String value;
 
-    public Name(String value) {
+    public Password(String value) {
         validate(value);
         this.value = value;
     }
@@ -15,7 +15,7 @@ public class Name {
 
     private void validateNull(String value) {
         if (value == null || value.isBlank()) {
-            throw new IllegalStateException("사용자 이름은 null 혹은 빈 문자열일 수 없습니다.");
+            throw new IllegalStateException("비밀번호는 null 혹은 빈 문자열일 수 없습니다.");
         }
     }
