+++ src/main/java/roomescape/model/member/Role.java
@@ -1,4 +1,4 @@
-package roomescape.model;
+package roomescape.model.member;
 
 import java.util.Arrays;
 import java.util.NoSuchElementException;
