+++ src/main/java/roomescape/repository/dao/JdbcMemberDao.java
@@ -3,8 +3,8 @@ package roomescape.repository.dao;
 import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.jdbc.core.RowMapper;
 import org.springframework.stereotype.Repository;
-import roomescape.model.Member;
-import roomescape.model.Role;
+import roomescape.model.member.Member;
+import roomescape.model.member.Role;
 
 import java.util.List;
 import java.util.Optional;
