+++ src/main/java/roomescape/service/AuthService.java
@@ -4,8 +4,8 @@ import io.jsonwebtoken.Claims;
 import io.jsonwebtoken.Jwts;
 import io.jsonwebtoken.security.Keys;
 import org.springframework.stereotype.Service;
-import roomescape.model.Member;
-import roomescape.model.Role;
+import roomescape.model.member.Member;
+import roomescape.model.member.Role;
 import roomescape.repository.MemberRepository;
 import roomescape.service.dto.AuthDto;
 import roomescape.service.dto.MemberInfo;
