+++ src/test/java/roomescape/model/MemberTest.java
@@ -4,6 +4,8 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.NullAndEmptySource;
 import org.junit.jupiter.params.provider.ValueSource;
+import roomescape.model.member.Member;
+import roomescape.model.member.Role;
 
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
