+++ src/test/java/roomescape/model/ReservationTest.java
@@ -5,6 +5,8 @@ import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
 import org.junit.jupiter.params.provider.ValueSource;
+import roomescape.model.member.Member;
+import roomescape.model.member.Role;
 import roomescape.service.dto.ReservationDto;
 
 import java.time.LocalDate;
