+++ src/test/java/roomescape/repository/ReservationRepositoryTest.java
@@ -8,6 +8,8 @@ import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.jdbc.core.simple.SimpleJdbcInsert;
 import roomescape.model.*;
+import roomescape.model.member.Member;
+import roomescape.model.member.Role;
 import roomescape.service.dto.ReservationDto;
 
 import java.time.LocalDate;
