+++ src/test/java/roomescape/service/AuthServiceTest.java
@@ -3,8 +3,8 @@ package roomescape.service;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import roomescape.model.Member;
-import roomescape.model.Role;
+import roomescape.model.member.Member;
+import roomescape.model.member.Role;
 import roomescape.repository.MemberRepository;
 import roomescape.repository.dao.MemberDao;
 import roomescape.service.dto.AuthDto;
